Add files via upload
A program that I wrote for my C++ class. First, it asks for a number of students. Then, according to the number that was entered, it asks for a grade of each student. The program organizes the grades and prints the final result, which displays the number of students for each grade category. There is also input validation implemented.